// RegexValidation/NTBRegexValidation/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    ValueToProcess: ComponentFramework.PropertyTypes.StringProperty;
    RegexExpression: ComponentFramework.PropertyTypes.StringProperty;
    UserNotificationText: ComponentFramework.PropertyTypes.StringProperty;
}
export interface IOutputs {
    ValueToProcess?: string;
}